Revert "订单id改为Long型"
This reverts commit 0edaf03de0f715a5e21a1fb2ab2e3850e51c1d0f.

+++ xinni-self-ordering-springboot3/pojo/src/main/java/com/xinni/entity/Order.java
@@ -21,7 +21,7 @@ public class Order implements Serializable {
     private static final long serialVersionUID = 1L;
 
     /**
-     * 订单状态 1待付款 2待接单 3已接单 4制作中 5已完成 6已取消
+     * 订单状态 1待付款 2待接单 3已接单 4派送中 5已完成 6已取消
      */
     public static final Integer PENDING_PAYMENT = 1;
     public static final Integer TO_BE_CONFIRMED = 2;
@@ -37,9 +37,9 @@ public class Order implements Serializable {
     public static final Integer PAID = 1;
     public static final Integer REFUND = 2;
 
-    private Long id;
+    private Integer id;
     private String number;  // 订单号
-    private Integer status; // 订单状态 1待付款 2待接单 3已接单 4制作中 5已完成 6已取消 7退款
+    private Integer status; // 订单状态 1待付款 2待接单 3已接单 4派送中 5已完成 6已取消 7退款
     private Integer userId; // 下单用户id
     private Integer addressBookId; // 地址id
     private LocalDateTime orderTime; // 下单时间
